// lib/gen/models/GetOrderItemResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MVC
 * MVC Assignment
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { ItemStatus } from './ItemStatus';
import { ItemStatusFromJSON, ItemStatusToJSON } from './ItemStatus';

/**
 *
 * @export
 * @interface GetOrderItemResponse
 */
export interface GetOrderItemResponse {
  /**
   *
   * @type {string}
   * @memberof GetOrderItemResponse
   */
  customInstructions?: string;
  /**
   *
   * @type {number}
   * @memberof GetOrderItemResponse
   */
  id?: number;
  /**
   *
   * @type {number}
   * @memberof GetOrderItemResponse
   */
  itemId?: number;
  /**
   *
   * @type {number}
   * @memberof GetOrderItemResponse
   */
  orderId?: number;
  /**
   *
   * @type {number}
   * @memberof GetOrderItemResponse
   */
  quantity?: number;
  /**
   *
   * @type {ItemStatus}
   * @memberof GetOrderItemResponse
   */
  status?: ItemStatus;
}

/**
 * Check if a given object implements the GetOrderItemResponse interface.
 */
export function instanceOfGetOrderItemResponse(value: object): value is GetOrderItemResponse {
  return true;
}

export function GetOrderItemResponseFromJSON(json: any): GetOrderItemResponse {
  return GetOrderItemResponseFromJSONTyped(json, false);
}

export function GetOrderItemResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetOrderItemResponse {
  if (json == null) {
    return json;
  }
  return {
    customInstructions: json['custom_instructions'] == null ? undefined : json['custom_instructions'],
    id: json['id'] == null ? undefined : json['id'],
    itemId: json['item_id'] == null ? undefined : json['item_id'],
    orderId: json['order_id'] == null ? undefined : json['order_id'],
    quantity: json['quantity'] == null ? undefined : json['quantity'],
    status: json['status'] == null ? undefined : ItemStatusFromJSON(json['status'])
  };
}

export function GetOrderItemResponseToJSON(json: any): GetOrderItemResponse {
  return GetOrderItemResponseToJSONTyped(json, false);
}

export function GetOrderItemResponseToJSONTyped(
  value?: GetOrderItemResponse | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    custom_instructions: value['customInstructions'],
    id: value['id'],
    item_id: value['itemId'],
    order_id: value['orderId'],
    quantity: value['quantity'],
    status: ItemStatusToJSON(value['status'])
  };
}