// lib/gen/models/GetOrderResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MVC
 * MVC Assignment
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import type { OrderStatus } from './OrderStatus';
import { OrderStatusFromJSON, OrderStatusToJSON } from './OrderStatus';

/**
 *
 * @export
 * @interface GetOrderResponse
 */
export interface GetOrderResponse {
  /**
   *
   * @type {number}
   * @memberof GetOrderResponse
   */
  customerId?: number;
  /**
   *
   * @type {number}
   * @memberof GetOrderResponse
   */
  id?: number;
  /**
   *
   * @type {string}
   * @memberof GetOrderResponse
   */
  orderedAt?: string;
  /**
   *
   * @type {OrderStatus}
   * @memberof GetOrderResponse
   */
  status?: OrderStatus;
  /**
   *
   * @type {number}
   * @memberof GetOrderResponse
   */
  tableNumber?: number;
}

/**
 * Check if a given object implements the GetOrderResponse interface.
 */
export function instanceOfGetOrderResponse(value: object): value is GetOrderResponse {
  return true;
}

export function GetOrderResponseFromJSON(json: any): GetOrderResponse {
  return GetOrderResponseFromJSONTyped(json, false);
}

export function GetOrderResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetOrderResponse {
  if (json == null) {
    return json;
  }
  return {
    customerId: json['customer_id'] == null ? undefined : json['customer_id'],
    id: json['id'] == null ? undefined : json['id'],
    orderedAt: json['ordered_at'] == null ? undefined : json['ordered_at'],
    status: json['status'] == null ? undefined : OrderStatusFromJSON(json['status']),
    tableNumber: json['table_number'] == null ? undefined : json['table_number']
  };
}

export function GetOrderResponseToJSON(json: any): GetOrderResponse {
  return GetOrderResponseToJSONTyped(json, false);
}

export function GetOrderResponseToJSONTyped(
  value?: GetOrderResponse | null,
  ignoreDiscriminator: boolean = false
): any {
  if (value == null) {
    return value;
  }

  return {
    customer_id: value['customerId'],
    id: value['id'],
    ordered_at: value['orderedAt'],
    status: OrderStatusToJSON(value['status']),
    table_number: value['tableNumber']
  };
}